checkbox: fix code formatting, gaining readability
The previous formatting was making the code unreadable, obey to the
common coding style instead.
While this changes a lot, there is still room for improvement.
It is a beginning and boosts the code to a readable level without
being pedantic.

started as a 5min job on the train, ended up in an hour job ;)
!Untested as I have no possibility to compile on my laptop atm!